AccuracyAssessment {greenbrown}	R Documentation

Accuracy assessment from a contingency table

Description

This function takes a contingency table as calculated with table or crosstab and computes an accuracy assessment, including the total accuracy, the user accuracy and the producer accuracy.

Usage

AccuracyAssessment(tab)

Arguments

tab
contingency table as calculated with table or crosstab

Value

The function returns the same frequency table as the input but with added row and column totals and total accuracy, user accuracy and producer accuracy.

Author(s)

Matthias Forkel <mforkel@bgc-jena.mpg.de>

References

Congalton, R.G. (1991): A review of assessing the accuracy of classifications of remotely sensed data. - Remote Sensing of Environment 1991, 37, 35–46.

See Also

CompareClassification, Kappa, TrendClassification

Examples



# two classifications: 
a <- c(1, 1, 1, 2, 2, 2, 3, 4, 5, 5, 5, 1, 1, 1, 2, 2, 2, 3, 4, 5, 5, 3, 3, 2, 2)
b <- c(1, 2, 1, 2, 2, 2, 3, 4, 2, 2, 5, 1, 2, 2, 2, 1, 2, 3, 4, 5, 5, 3, 3, 2, 2)

# calculate first a contingency table
tab <- table(a, b)

# calculate now the accuracy assessment
AccuracyAssessment(tab)

# calculate Kappa coeffcient
Kappa(tab)


[Package greenbrown version 2.1 Index]