AllTsteps {greenbrown}	R Documentation

Convert an irregular zoo time series to a zoo time series with all time steps

Description

Irregular 'zoo' time series with missing time steps are converted to a 'zoo' time series including all time steps. Observations at time steps that were missing in the original time series are filled with NA.

Usage

AllTsteps(x, by = "day", start.jan = FALSE, end.dec = FALSE, 
    exclude.feb29 = FALSE, ...)

Arguments

x
a time series of class 'zoo'

by
time step of full time series, e.g. "day", "month"

start.jan
Should the full time series start in January? If FALSE, the full time series will start at the first observation in x.

end.dec
Should the full time series end in December? If FALSE, the full time series will end at the last observation in x.

exclude.feb29
Should 29th Februaries be excluded in case of a daily time series (i.e. if by = 'day')?

...
further arguments (currently not used)

Author(s)

Matthias Forkel <matthias.forkel@geo.tuwien.ac.at> [aut, cre]

Examples

x <- zoo(rnorm(5), as.Date(c("2010-01-15", "2010-02-15", "2010-07-15", 
 "2010-08-15", "2010-09-15")))
x
AllTsteps(x, by="month")

[Package greenbrown version 2.4.3 Index]