Decompose {greenbrown}	R Documentation

Simple decomposition of time series

Description

This function decomposes time series in different components using a simple step-wise approach.

Usage

Decompose(Yt, breaks = 0, mosum.pval = 0.05)

Arguments

Yt
univariate time series of class ts

breaks
maximal number of breaks in the trend component to be calculated (integer number).

mosum.pval
Maximum p-value for the OLS-MOSUM test in order to search for breakpoints. If p = 0.05, breakpoints will be only searched in the time series trend component if the OLS-MOSUM test indicates a significant structural change in the time series. If p = 1 breakpoints will be always searched regardless if there is a significant structural change in the time series or not.

Details

The decomposition of the time series is based on a simple step-wise approach:

The mean of the NDVI time series is calculated.

In the second step, monthly values are aggregated per year by using the average value and the trend is calculated based on annual aggregated values using TrendAAT.

The mean of the time series and the derived trend component from step (2) are subtracted from the annual values to derive the trend-removed and mean-centred annual values (annual anomalies). If the trend slope is not significant (p > 0.05), only the mean is subtracted.

In the next step, the mean, the trend component and the annual anomalies are subtracted from the original time series to calculate a detrended, mean-centered and for annual anomalies adjusted time series. From this time series the seasonal cycle is estimated as the mean seasonal cycle.

In the last step, the short-term anomalies are computed. For this, the mean, the trend component, the annual anomalies and the mean seasonal cycle are subtracted from the original time series.

Value

The function returns a multi-variate object of class ts including the time series components.

Author(s)

Matthias Forkel <mforkel@bgc-jena.mpg.de> [aut, cre]

References

Forkel, M., N. Carvalhais, J. Verbesselt, M. Mahecha, C. Neigh and M. Reichstein (2013): Trend Change Detection in NDVI Time Series: Effects of Inter-Annual Variability and Methodology. - Remote Sensing 5.

See Also

GetTsStatisticsRaster

Examples


# load a time series of Normalized Difference Vegetation Index
data(ndvi)
plot(ndvi)

# decompose this time series
ndvi.dc <- Decompose(ndvi)
plot(ndvi.dc)
	
ndvi.dc2 <- Decompose(ndvi, breaks=2, mosum.pval=1)
plot(ndvi.dc2)


[Package greenbrown version 2.2 Index]