GetTsStatisticsRaster {greenbrown}	R Documentation

Estimate statistical properties of time series in a multi-temporal raster dataset

Description

This function computes statistical properties of the time series in a multi-temporal raster dataset. It calls Decompose to decompose the time series of each grid cell of a raster brick into a trend, inter-annual variability, seasonal and short-term variability time series components. In a next step the mean, the trend slope, the range and standard deviation of the inter-annual variability, the range of the seasonal cycle as well as the range and standard devaition of the short-term variability are calculated.

Usage

GetTsStatisticsRaster(r, start = c(1982, 1), freq = 12)

Arguments

r
object of class brick with multi-temporal data.

start
first time step, e.g. c(1982, 1) for January 1982. See ts for details.

freq
the number of observations per unit of time, e.g. 12 for monthly data or 24 for bi-monthly data. See ts for details.

Value

The function returns a RasterBrick with 7 layers: mean, trend slope, range of inter-annual variabililty, standard deviation of inter-annual variabililty, range of seasonal cycle, range and standard deviation of short-term variabililty.

Author(s)

Matthias Forkel <mforkel@bgc-jena.mpg.de> [aut, cre]

See Also

Decompose

Examples


# load a multi-temporal raster dataset of Normalized Difference Vegetation Index
data(ndvimap)
plot(ndvimap, 8)

# calculate time series statistics 
ndvimap.tsstat <- GetTsStatisticsRaster(ndvimap)
plot(ndvimap.tsstat)

[Package greenbrown version 2.2 Index]