KGETrendUncertainty {greenbrown}	R Documentation

Compute uncertainty of Kling-Gupta efficiency based on beginning and end of time series

Description

This function samples time series for different combinations of start and end years and computes for each combination the KGE (see KGE).

Usage

KGETrendUncertainty(x, ref, trend = TRUE, eTrend.ifsignif = FALSE, 
    sample.method = c("sample", "all", "none"), sample.min.length = 0.75, 
    sample.size = 30, ...)

Arguments

x
time series from model result or factorial model run

ref
reference time series (observation or standard model run)

trend
Include the effect of trend in the calculation?

eTrend.ifsignif
compute effect on trend only if trend in reference series is significant, if FALSE compute always effect on trend (if trend = TRUE)

sample.method
Sampling method for combinations of start and end dates to compute uncertainties in trends. If "sample" (default), trend statistics are computed for a sample of combinations of start and end dates according to sample.size. If "all", trend statistics are computed for all combinations of start and end dates longer than sample.min.length. If "none", trend statistics will be only computed for the entire time series (i.e. no sampling of different start and end dates).

sample.min.length
Minimum length of the time series (as a fraction of total length) that should be used to compute trend statistics. Time windows between start and end that are shorter than min.length will be not used for trend computation.

sample.size
sample size (number of combinations of start and end dates) to be used if method is sample.

...
further arguments for the function Trend

Details

...

Value

The function returns a data.frame with the following components:

start start of the time series

end end of the time series

length length of the time series

KGE Kling-Gupta effciency = 1 - eTotal

eTotal total effect, i.e. euclidean distance

fMean fraction of mean to the total effect

fVar fraction of variance to the total effect

fCor fraction of correlation to the total effect

fTrend fraction of trend to the total effect (only if trend=TRUE)

eMean effect of mean

eVar effect of variance

eCor effect of correlation

eTrend effect of trend (only if trend=TRUE)

Author(s)

Matthias Forkel <mforkel@bgc-jena.mpg.de> [aut, cre]

References

Gupta, H. V., H. Kling, K. K. Yilmaz, G. F. Martinez (2009): Decomposition of the mean squared error and NSE performance criteria: Implications for improving hydrological modelling. Journal of Hydrology 377, 80-91.

See Also

Trend

Examples


# load a time series of NDVI (normalized difference vegetation index)
data(ndvi)
plot(ndvi)

# change the variance and compute effect
x <- ndvi + rnorm(length(ndvi), 0, 0.01)
plot(x, ndvi); abline(0,1)
unc <- KGETrendUncertainty(x, ndvi)
hist(unc$KGE)
	

[Package greenbrown version 2.2 Index]