Kappa {greenbrown}	R Documentation

Calculate the Kappa coefficient of two classifications

Description

This function takes a contingency table as calculated with table or crosstab and computes the Kappa coefficient.

Usage

Kappa(tab)

Arguments

tab
contingency table as calculated with table or crosstab

Value

Kappa coeffcient

Author(s)

Matthias Forkel <matthias.forkel@geo.tuwien.ac.at> [aut, cre]

References

Congalton, R.G. (1991): A review of assessing the accuracy of classifications of remotely sensed data. - Remote Sensing of Environment 1991, 37, 35-46.

See Also

CompareClassification, AccuracyAssessment, TrendClassification

Examples

# two classifications: 
a <- c(1, 1, 1, 2, 2, 2, 3, 4, 5, 5, 5, 1, 1, 1, 2, 2, 2, 3, 4, 5, 5, 3, 3, 2, 2)
b <- c(1, 2, 1, 2, 2, 2, 3, 4, 2, 2, 5, 1, 2, 2, 2, 1, 2, 3, 4, 5, 5, 3, 3, 2, 2)

# calculate first a contingency table
tab <- table(a, b)

# calculate now the accuracy assessment
AccuracyAssessment(tab)

# calculate Kappa coeffcient
Kappa(tab)

[Package greenbrown version 2.4.3 Index]