NamesPhenologyRaster {greenbrown}	R Documentation

Get the layer names for a PhenologyRaster raster brick

Description

This function returns the layer names of a raster brick that was created using PhenologyRaster

Usage

NamesPhenologyRaster(x, start = NULL)

Arguments

x
RasterBrick as created with PhenologyRaster or integer as the number of years of the input data when the function PhenologyRaster was called.

start
beginning of the time series.

Author(s)

Matthias Forkel <mforkel@bgc-jena.mpg.de>

See Also

PhenologyRaster

Examples


# load a multi-temporal raster dataset of Normalized Difference Vegetation Index
data(ndvimap)
plot(ndvimap, 8)

# calculate phenology
phenmap <- PhenologyRaster(ndvimap, start=c(1982, 1), freq=12)
plot(phenmap)
plot(phenmap)

# layer names of the result
NamesPhenologyRaster(30)
NamesPhenologyRaster(phenmap)
NamesPhenologyRaster(phenmap, start=1982)
names(phenmap)



[Package greenbrown version 2.2 Index]