NoBP {greenbrown}	R Documentation

Initialize an empty list with breakpoints

Description

This is an internal function to make an empty list of breakpoints. For the user there is usually no need to use this function.

Usage

NoBP()

Value

empty list with slot 'breakpoints'

Author(s)

Matthias Forkel <mforkel@bgc-jena.mpg.de>

[Package greenbrown version 2.2 Index]