NoTrend {greenbrown}	R Documentation

Initialize an empty object of class "Trend"

Description

This is an internal function to make an empty list of class Trend. For the user there is usually no need to use this function.

Usage

NoTrend(Yt)

Arguments

Yt
univariate time series of class ts

Value

The function returns a list of class "Trend".

Author(s)

Matthias Forkel <mforkel@bgc-jena.mpg.de>

[Package greenbrown version 2.1 Index]