PhenoDeriv {greenbrown}	R Documentation

Method 'Deriv' to calculate phenology metrics

Description

This function implements the derivative method for phenology. This is rather an internal function; please use the function Phenology to apply this method.

Usage

PhenoDeriv(x, min.mean = 0.1, calc.pheno = TRUE, plot = FALSE, 
    ...)

Arguments

x
seasonal cycle of one year

min.mean
minimum mean annual value in order to calculate phenology metrics. Use this threshold to suppress the calculation of metrics in grid cells with low average values

calc.pheno
calculate phenology metrics or return NA?

plot
plot results?

...
further arguments (currently not used)

Value

The function returns a vector with SOS, EOS, LOS, POP, MGS, RSP, RAU, PEAK, MSP and MAU.

Author(s)

Matthias Forkel <matthias.forkel@geo.tuwien.ac.at> [aut, cre]

See Also

Phenology

Examples

data(ndvi)
plot(ndvi)

# perform time series preprocessing from first year of data
x <- TsPP(ndvi, interpolate=TRUE)[1:365]
plot(x)

# calculate phenology metrics for first year
PhenoDeriv(x, plot=TRUE)

[Package greenbrown version 2.4.2 Index]