PhenopixMY {greenbrown}	R Documentation

Multi-year phenology analysis using phenopix

Description

This function takes a multi-year time series and applies curve fitting and phenology extraction functions based on the greenProcess function in the phenopix package. The function returns an object of class PhenopixMY (phenopix multi-year) which contains a list of phenopix objects. PhenopixMY can be plotted using plot.PhenopixMY.

Usage

PhenopixMY(ts, fit, threshold = NULL, plot = FALSE, ...)

Arguments

ts
a time series of class 'ts' or 'zoo' with multiple years of data

fit
fitting function to be applied, available options are: spline, beck, elmore, klosterman, gu (see greenProcess)

threshold
threshold to be applied to compute phenology metrics, available options are: trs, derivatives, klosterman, gu (see greenProcess)

plot
plot phenopix object of each year, using plot.phenopix

...
further arguments as in greenProcess

Value

An object of class phenopixmy with dedicated functions: plot(), print(). The structure is actually a list.

Author(s)

Matthias Forkel <matthias.forkel@geo.tuwien.ac.at> [aut, cre]

References

Filippa, G., Cremonese, E., Migliavacca, M., Galvagno, M., Forkel, M., Wingate, L., Tomelleri, E., Morra di Cella, U. and Richardson, A. D.: Phenopix: A R package for image-based vegetation phenology, Agricultural and Forest Meteorology, 220, 141-150, doi:10.1016/j.agrformet.2016.01.006, 2016.

See Also

greenProcess, plot.PhenopixMY, Phenology

Examples

data(ndvi)
plot(ndvi)

ppixmy <- PhenopixMY(ndvi, "spline", "trs")
plot(ppixmy)

plot(ppixmy, type="metrics")

[Package greenbrown version 2.4.2 Index]