PlotPhenCycle {greenbrown}	R Documentation

Plot a easonal cycle with phenology metrics

Description

This function plots a seasonal cycle with phenology metrics.

Usage

PlotPhenCycle(x, xpred = NULL, metrics, xlab = "DOY", ylab = "NDVI", 
    trs = NULL, main = "", ...)

Arguments

x
values of one year

xpred
smoothed/predicted values

metrics
vector of pheology metrics

xlab
label for x-axis

ylab
label for y-axis

trs
threshold for threshold methods

main
title

...
further arguments (currently not used)

Author(s)

Matthias Forkel <mforkel@bgc-jena.mpg.de> [aut, cre]

See Also

Phenology

Examples



data(ndvi)
plot(ndvi)

# perform smoothing, gap filling and interpolation and extract seasonal cycle from first year
x <- TsPP(ndvi, interpolate=TRUE)[1:365]
plot(x)

# calculate phenology metrics for first year
metrics <- PhenoTrs(x, approach="White")
PlotPhenCycle(x, metrics=metrics)


[Package greenbrown version 2.2 Index]