ReadVI3g {greenbrown}	R Documentation

Read and pre-process GIMMS VI3g binary files

Description

This function reads GIMMS VI3g binary files, pre-processes the values (exclusion of flagged values, subset for area of interest) and returns the result as a raster layer. The function can be used to read the original GIMMS NDVI3g data files to R.

Usage

ReadVI3g(file, flag = 2:7, ext = c(-180, 180, -90, 90))

Arguments

file
GIMMS VI3g file name

flag
vector of quality flags that should be excluded. Default: 2:7 (all values with reduced quality excluded). If you want to keep all values set flag=NA.

ext
extent (xmin, xmax, ymin, ymax) for which to extract the data

Details

The GIMMS NDVI3g dataset comes with the following quality flags:

FLAG = 1 (Good value)

FLAG = 2 (Good value, possibly snow)

FLAG = 3 (NDVI retrieved from spline interpolation)

FLAG = 4 (NDVI retrieved from spline interpolation, possibly snow)

FLAG = 5 (NDVI retrieved from average seasonal profile)

FLAG = 6 (NDVI retrieved from average seasonal profile, possibly snow)

FLAG = 7 (missing data)

Value

raster layer

Author(s)

Matthias Forkel <mforkel@bgc-jena.mpg.de>

Examples


#	data <- ReadVI3g("geo00oct15a.n14-VI3g")
#	plot(data)

[Package greenbrown version 2.1 Index]