SSASeasonalCycle {greenbrown}	R Documentation

Calculate a modulated seasonal cycle of a time series based on singular spectrum analysis (SSA)

Description

The function calculates a seasonal cycle based on 1-D singular spectrum analysis (Golyandina et al. 2001) as implemented in the Rssa package. See ssa for details.

Usage

SSASeasonalCycle(ts)

Arguments

ts
univariate time series of class ts

Author(s)

Matthias Forkel <matthias.forkel@geo.tuwien.ac.at> [aut, cre]

References

Golyandina, N., Nekrutkin, V. and Zhigljavsky, A. (2001): Analysis of Time Series Structure: SSA and related techniques. Chapman and Hall/CRC. ISBN 1584881941

See Also

TrendSeasonalAdjusted, ssa, reconstruct

Examples

## load a time series of Normalized Difference Vegetation Index
#data(ndvi)
#plot(ndvi)

## estimate the seasonal cycle using SSA
#ndvi.cycle <- SSASeasonalCycle(ndvi)
#plot(ndvi.cycle)

## the mean seasonal cycle is centered to 0,
## add the mean of the time series if you want to overlay it with the original data
#plot(ndvi)
#lines(ndvi.cycle + mean(ndvi, na.rm=TRUE), col="blue")
	

[Package greenbrown version 2.4.3 Index]