SimIAV {greenbrown}	R Documentation

Simulate the inter-annual variability component of a surrogate time series

Description

The function simulates the inter-annual variability component of a time series based on normal-distributed random values.

Usage

SimIAV(sd = 0.015, range = sd * 2, nyears = 30, start = c(1982, 
    1), freq = 12)

Arguments

sd
standard deviation of the annual mean values

range
range of the annual mean values

nyears
number of years

start
beginning of the time series (i.e. the time of the first observation). The default is c(1982, 1), i.e. January 1982 which is the usual start date to compute trends on long-term series of satellite observations of NDVI. See ts for further examples.

freq
The frequency of observations. The default is 12 for monthly observations. Use 24 for bi-monthly observations, 365 for daily observations or 1 for annual observations. See ts for further examples.

Value

time series of class ts

Author(s)

Matthias Forkel <mforkel@bgc-jena.mpg.de>

See Also

SimTs

Examples


It <- SimIAV(sd=0.015, range=0.05, nyears=30)
plot(It)

[Package greenbrown version 2.1 Index]