SimTrend {greenbrown}	R Documentation

Simulate trend and breakpoints of a surrogate time series

Description

The function simulates the trend component with breakpoints of a time series.

Usage

SimTrend(slope = c(0.002, -0.004), breaks = 165, abrupt = TRUE, 
    n = 360, start = c(1982, 1), freq = 12)

Arguments

slope
slope of the trend in each time series segment. slope should be a numeric vector. The length of this vector determines the number of segments.

breaks
position of the breakpoints in the time series. You should specify one more slope than breakpoint.

abrupt
Should the trend at the breakpoints change abrupt (TRUE) or gradual (FALSE)?

n
length of the time series

start
beginning of the time series (i.e. the time of the first observation). The default is c(1982, 1), i.e. January 1982 which is the usual start date to compute trends on long-term series of satellite observations of NDVI. See ts for further examples.

freq
The frequency of observations. The default is 12 for monthly observations. Use 24 for bi-monthly observations, 365 for daily observations or 1 for annual observations. See ts for further examples.

Value

time series of class ts

Author(s)

Matthias Forkel <matthias.forkel@geo.tuwien.ac.at> [aut, cre]

See Also

SimTs

Examples

Tt <- SimTrend(slope=c(0.003, -0.001), breaks=150)
plot(Tt)

[Package greenbrown version 2.4.3 Index]