TrendLongestSEG {greenbrown}	R Documentation

Extract slope and p-value for the longest time series segment from a TrendRaster raster brick

Description

This function extracts the slope and p-value of a trend for the longest time series segment from a raster brick that was created with TrendRaster

Usage

TrendLongestSEG(r)

Arguments

r
RasterBrick as created with TrendRaster or object of class 'Trend' as returned by Trend

Author(s)

Matthias Forkel <matthias.forkel@geo.tuwien.ac.at> [aut, cre]

See Also

TrendRaster

Examples

# load a multi-temporal raster dataset of Normalized Difference Vegetation Index
data(ndvimap)
plot(ndvimap, 8)

# calculate trend 
trendmap <- TrendRaster(ndvimap, start=c(1982, 1), freq=12, method="AAT", breaks=1)
plot(trendmap)

# select trend and p-value only for the longest time series segment
trendmap.longestseg <- TrendLongestSEG(trendmap)
plot(trendmap.longestseg)

[Package greenbrown version 2.4.3 Index]