WriteNCDF {greenbrown}	R Documentation

Write raster objects to NetCDF files

Description

This function writes raster layers to NetCDF files including meta information as variable names and units and time axes.

Usage

WriteNCDF(data = NULL, var.name, var.unit, var.longname = "", 
    file = NULL, time = NULL, layernames = NULL, data.name = NA, 
    region.name = NA, file.title = var.longname, file.description = NULL, 
    reference = "", provider = "", creator = "greenbrown R package", 
    naflag = -9999, scale = 1, offset = 0, overwrite = FALSE)

Arguments

data
raster layer or raster brick

var.name
variable name

var.unit
variable unit

var.longname
variable long name

file
file name. If NULL the file name will be created from the variable name and the dimensions of the data.

time
vector of time steps for each layer.

layernames
layer names if layers are not time steps.

data.name
name of the dataset

region.name
name of the region

file.title
title of the file

file.description
description of the file

reference
reference for the dataset

provider
dataset provider

creator
dataset creator

naflag
flag for NA values

scale
sclaing values for the data

offset
offset value

overwrite
overwrite existing file?

Author(s)

Matthias Forkel <mforkel@bgc-jena.mpg.de> [aut, cre]

[Package greenbrown version 2.2 Index]