WriteNCDF4 {greenbrown}	R Documentation

Write NetCDF files

Description

Writes NetCDF files from rasters and makes sure that meta-information is properly defined.

Usage

WriteNCDF4(data.l, var.name, var.unit, time = as.Date("2000-01-01"), 
    var.description = var.name, file = NULL, data.name = NA, 
    region.name = NA, file.title = var.name, file.description = var.name, 
    reference = "", provider = "", creator = "", missval = -9999, 
    scale = 1, offset = 0, compression = 9, overwrite = FALSE)

Arguments

data.l
a single Raster* object or a list of Raster* objects

var.name
vector of variable names

var.unit
vector of variable units

time
vector of time steps for each layer.

var.description
vector of variable descriptions

file
file name. If NULL the file name will be created from the variable name and the dimensions of the data.

data.name
name of the dataset

region.name
name of the region

file.title
title of the file

file.description
description of the file

reference
reference for the dataset

provider
dataset provider

creator
dataset creator

missval
flag for missing/NA values

scale
scaling values for the data

offset
offset value

compression
If set to an integer between 1 (least compression) and 9 (most compression), this enables compression for the variable as it is written to the file. Turning compression on forces the created file to be in netcdf version 4 format, which will not be compatible with older software that only reads netcdf version 3 files.

overwrite
overwrite existing file?

Author(s)

Matthias Forkel <matthias.forkel@geo.tuwien.ac.at> [aut, cre]

[Package greenbrown version 2.4.2 Index]