greenbrown-package {greenbrown}	R Documentation

greenbrown - land surface phenology and trend analysis

Description

greenbrown is a collection of functions to analyse trends, trend changes and phenology events in gridded time series like from satellite observations or climate model simulations. The package provides access to different methods for 1) trend and breakpoint analysis, 2) time series smoothing and interpolation, and 3) analysis of land surface phenology.

Author(s)

Matthias Forkel <mforkel@bgc-jena.mpg.de>

[Package greenbrown version 2.2 Index]