plot.Phenology {greenbrown}	R Documentation

Plot time series of phenology metrics

Description

This is the standard plot function for results of the Phenology function. See plot.default for further specifications of basic plots.

Usage

## S3 method for class 'Phenology'
plot(x, type = c("sos", "eos", "pop"), ylab = NULL, 
    ylim = NULL, add = FALSE, col = "black", add.trend = TRUE, 
    ...)

Arguments

x
Object of class 'Phenology' as returned from function Phenology

type
varaible names that should be plotted from the Phenology object

ylab
a title for the y axis

ylim
limits for y-axis

add
add time series to exisiting plot?

col
line colors

add.trend
add trend lines to phenology time series?

...
Further arguments that can be passed plot.default

Author(s)

Matthias Forkel <mforkel@bgc-jena.mpg.de>

See Also

plot.default, plot.ts

Examples


# load a time series of NDVI (normalized difference vegetation index)
data(ndvi)
plot(ndvi)
	
# calculate phenology metrics
phen <- Phenology(ndvi)
phen

# plot phenology metrics
plot(phen)
	

[Package greenbrown version 2.2 Index]