plot.Trend {greenbrown}	R Documentation

Plot trend and breakpoint results

Description

This is the standard plot function for results of the Trend function. See plot.default for further specifications of basic plots.

Usage

## S3 method for class 'Trend'
plot(x, ylab = "NDVI", add = FALSE, col = c("black", "blue", 
    "red", "blue"), lty = c(2, 1, 2, 3), lwd = 1, symbolic = TRUE, 
    legend = FALSE, axes = TRUE, ...)

Arguments

x
Object of class 'Trend' as returned from function Trend

ylab
A title for the y axis

add
add to exisiting plot

col
colors for (1) time series, (2) trend line, (3) breakpoints and (4) trend uncertainty

lty
line types for (1) time series, (2) trend line, (3) breakpoints and (4) trend uncertainty

lwd

symbolic
add significance as symbols (TRUE). If TRUE the p-value of a trend slope is added as symbol as following: *** (p <= 0.001), ** (p <= 0.01), * (p <= 0.05), . (p <= 0.1) and no symbol if p > 0.1.

legend
add slope and p-value as legend

axes
plot axes?

...
Further arguments that can be passed plot.default

Author(s)

Matthias Forkel <matthias.forkel@geo.tuwien.ac.at> [aut, cre]

See Also

plot.default, plot.ts

Examples

# load a time series of Normalized Difference Vegetation Index
data(ndvi)
plot(ndvi)

# calculate a trend and look at the results
ndvi.trend <- Trend(ndvi)
ndvi.trend
plot(ndvi.trend)

ndvi.trend.aat <- Trend(ndvi, method="AAT", mosum.pval=1)
plot(ndvi.trend.aat)
plot(ndvi.trend.aat, symbolic=FALSE)
plot(ndvi.trend.aat, symbolic=FALSE, legend=TRUE)

ndvi.trend.stm <- Trend(ndvi, method="STM", mosum.pval=1)
plot(ndvi.trend.stm)

plot(ndvi.trend.aat, symbolic=TRUE, ylim=c(0.23, 0.31), 
   col=c("blue", "blue", "red"))
plot(ndvi.trend.stm, symbolic=TRUE, col=c("darkgreen", "darkgreen", "red"), 
	lty=c(0, 1, 1), add=TRUE)
	

[Package greenbrown version 2.4.3 Index]