summary.Trend {greenbrown}	R Documentation

Print a summary of calculated trends

Description

The function prints a summary the results from Trend.

Usage

## S3 method for class 'Trend'
summary(x)

Arguments

x
Object of class 'Trend' as returned from function Trend

Author(s)

Matthias Forkel <mforkel@bgc-jena.mpg.de> [aut, cre]

Examples


# load a time series of Normalized Difference Vegetation Index
data(ndvi)
plot(ndvi)

# calculate a trend and look at the results
ndvi.trend <- Trend(ndvi)
ndvi.trend
print(ndvi.trend)
summary(ndvi.trend)
	

[Package greenbrown version 2.2 Index]