greenbrown-package {greenbrown}	R Documentation

Land Surface Phenology and Trend Analysis

Description

Collection of functions to analyse trends, trend changes and phenology events in gridded time series like from satellite observations or climate model simulations. The package provides access to different methods for 1) trend and breakpoint analysis, 2) time series smoothing and interpolation, and 3) analysis of land surface phenology.

Details

Package: 	 greenbrown

Type: 	 Package

Version: 	 2.4.3

Date: 	 2015-11-17

License: 	 GPL-2

Satellite observations are used to monitor temporal changes of the terrestrial vegetation. Satellite-derived time series of vegetation indices such as Normalized Difference Vegetation Index (NDVI) are indicative of the coverage of green vegetation, photosynthetic activity and green biomass. However, the analysis of vegetation index time series is often dependent on the used analysis methods. The package provides access to different methods for 1) trend and breakpoint analysis, 2) time series smoothing and interpolation, and 3) analysis of land surface phenology.

The methods for trend and breakpoint analysis mostly refer to Forkel et al. (2013). For phenology methods refer to Forkel et al. (2015).

Author(s)

Matthias Forkel <matthias.forkel@geo.tuwien.ac.at> [aut, cre]

Maintainer: Matthias Forkel <matthias.forkel@geo.tuwien.ac.at>, Thomas Wutzler <twutz@bgc-jena.mpg.de>

References

Forkel, M., Carvalhais, N., Verbesselt, J., Mahecha, M., Neigh, C., Reichstein, M., 2013. Trend Change Detection in NDVI Time Series: Effects of Inter-Annual Variability and Methodology. Remote Sensing 5, 2113-2144. doi:10.3390/rs5052113

Forkel, M., Migliavacca, M., Thonicke, K., Reichstein, M., Schaphoff, S., Weber, U., Carvalhais, N., 2015. Codominant water control on global interannual variability and trends in land surface phenology and greenness. Glob Change Biol 21, 3414–3435. doi:10.1111/gcb.12950

See Also

http://greenbrown.r-forge.r-project.org/

[Package greenbrown version 2.4.3 Index]